+++ src/index.js
@@ -1,1 +1,7 @@
+import SortingView from './sorting/sorting-view';
 import './styles/global.css';
+
+
+let main = document.querySelector('main');
+
+main.replaceWith(new SortingView().getView());