+++ CHANGELOG.md
@@ -5,6 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.0.3] - 2022-07-16
+### Added
+- added bubble sort
+- added insertion sort
+- added bubble sort
+- functionality to sorting alogrithms view
+
+## [0.0.2] - 2022-07-16
+### Added
+- sorting algorithm visualization structure
+- sorting-view.js
+- array-bars.js
+- sorting-view.css
+
 ## [0.0.1] - 2022-08-15
 ### Added
 - initial structure